+++ commands/Set.go
@@ -33,7 +33,7 @@ func Set(client *discordgo.Session, interaction *discordgo.Interaction) error {
 		client.InteractionRespond(interaction, &discordgo.InteractionResponse{
 			Type: discordgo.InteractionResponseChannelMessageWithSource,
 			Data: &discordgo.InteractionResponseData{
-				Flags: uint64(discordgo.MessageFlagsEphemeral),
+				Flags: discordgo.MessageFlagsEphemeral,
 				Embeds: []*discordgo.MessageEmbed{
 					{
 						Author: &discordgo.MessageEmbedAuthor{
@@ -74,7 +74,7 @@ func Set(client *discordgo.Session, interaction *discordgo.Interaction) error {
 		client.InteractionRespond(interaction, &discordgo.InteractionResponse{
 			Type: discordgo.InteractionResponseChannelMessageWithSource,
 			Data: &discordgo.InteractionResponseData{
-				Flags: uint64(discordgo.MessageFlagsEphemeral),
+				Flags: discordgo.MessageFlagsEphemeral,
 				Embeds: []*discordgo.MessageEmbed{
 					{
 						Author: &discordgo.MessageEmbedAuthor{
@@ -95,7 +95,7 @@ func Set(client *discordgo.Session, interaction *discordgo.Interaction) error {
 	client.InteractionRespond(interaction, &discordgo.InteractionResponse{
 		Type: discordgo.InteractionResponseChannelMessageWithSource,
 		Data: &discordgo.InteractionResponseData{
-			Flags: uint64(discordgo.MessageFlagsEphemeral),
+			Flags: discordgo.MessageFlagsEphemeral,
 			Embeds: []*discordgo.MessageEmbed{
 				{
 					Author: &discordgo.MessageEmbedAuthor{
