+++ commands/Settings.go
@@ -35,7 +35,7 @@ func Settings(client *discordgo.Session, interaction *discordgo.Interaction) err
 	client.InteractionRespond(interaction, &discordgo.InteractionResponse{
 		Type: discordgo.InteractionResponseChannelMessageWithSource,
 		Data: &discordgo.InteractionResponseData{
-			Flags: uint64(discordgo.MessageFlagsEphemeral),
+			Flags: discordgo.MessageFlagsEphemeral,
 			Embeds: []*discordgo.MessageEmbed{
 				{
 					Author: &discordgo.MessageEmbedAuthor{
